omx_vdec: if using AndroidNativeBuffer, map OMX color format to HAL
If the use of ANB was enabled by stagefright, we have to report a
HAL format in GetParameter(PortDefinition). However, the actual port
definition stores an OMX format. Also, lot's of other places in this
OMX implementation rely on everything being in OMX color format. So,
we don't actually change the internal format info to a HAL format but
rather fake a HAL format in the return value of the GetParameter call.

Change-Id: I24afb4dd4275b8594eb24d473fde6c35e57cbd54
Signed-off-by: Dima Zavin <dima@android.com>