HACK: exynos5: gscaler: do not skip s_fmt/reqbufs on gsc_convert path
This fixes the bug where still capture always captures the same
image and video recording only taking first image.

The issue is that reqbufs is reset to 0 and fmt is not set
when the buffers try to be enqueued and stremaing turned on in the path
from csc_convert().

This is temporary as the gscaler code really needs a huge cleanup.

Bug: 6925368
Change-Id: Ia8e3aee0e0b50c6a69e1eda219e484713088c042
Signed-off-by: Dima Zavin <dima@android.com>