exynos_omx: multi_thread: Change bitrate setting in encode component.
Change bitrate setting in encode component for VBR and CBR support.
CBR mains a relative constant bit rate for a small window of time period and the actual bit rate for that small window should be really close to the target bit rate.
VBR maintains a larger variable bit rate for a small window of time duration, but the avg bit rate still needs to converge to the target bit rate for a relatively large window of time duration.

Change-Id: I25e7ccdf3c53e6abacc1c7ed9235292ab97b940a
Signed-off-by: SeungBeom Kim <sbcrux.kim@samsung.com>