Don't call my-dir
Because this is a "config" makefile which will be included in the middle
of another Android.mk, while the build system depends on my-dir to set
up LOCAL_MODULE_MAKEFILE, so the value of LOCAL_MODULE_MAKEFILE is
changed from the point where mobicore/common/LogWrapper/Android.mk is
included. That leads to incorrect dependency of export_includes.

Change-Id: I6af1e1cf4574d9ed15bf6d82bdf6a0815fb667ce