exynos: gralloc: register buffers in alloc
gralloc_alloc was creating buffers with ion_alloc_fd, which results
in a dmabuf but no ion handle.  This makes it hard to track buffers
that are never registered with another process, or ones that are
unregistered from their process like the wallpaper.

Call gralloc_register_buffer from inside alloc to convert the dmabufs
to ion handles, and gralloc_unregister_buffer inside free.
gralloc_register_buffer currently unnecissarily maps the buffer as
well, which gralloc_lock will do as necessary, so remove the map.
gralloc_unregister_buffer will unmap if necesssary, so remove
the call to grallocUnmap in gralloc_free.  That leaves no callers
to grallocUnmap (and there weren't any callers to grallocMap), so
remove them and gr.h.

Change-Id: I20cf2d50732d018f12779d78582686ca622a332a